add EGO application
