+++ doc/_src_docs/applications.rstx
@@ -11,6 +11,6 @@ Three methods are available:
    applications/moe
    applications/vfm
    applications/mfk
-
+   applications/ego
 
 The intent is to provide applications of surrogate models in higher level methods.
