+++ doc/_src_docs/applications/mfk.rst
@@ -103,7 +103,7 @@ Usage
    Training
      
      Training ...
-     Training - done. Time (sec):  0.0156002
+     Training - done. Time (sec):  0.0155997
   ___________________________________________________________________________
      
    Evaluation
