+++ doc/_src_docs/applications/moe.rst
@@ -138,54 +138,54 @@ Usage
   
 ::
 
-  Kriging 1.418429438614483
+  Kriging 1.4184483897492126
   LS 1.4418735729979624
   QP 1.4542641515660801
-  KPLS 1.4186236789470246
-  KPLSK 1.4184385612110044
-  RBF 1.4178300357368523
-  RMTC 1.430746629259529
-  RMTB 1.4285575959464327
+  KPLS 1.4186236791852722
+  KPLSK 1.4184385611898738
+  RBF 1.4178300358901577
+  RMTC 1.4314324815657455
+  RMTB 1.4285459482856704
   IDW 1.4188134912804515
   Best expert = RBF
-  Kriging 1.0172805901833657
-  LS 1.0451612679414988
+  Kriging 1.0172805902016249
+  LS 1.045161267941499
   QP 1.021870693299392
-  KPLS 1.0198972713675998
-  KPLSK 1.016469984701472
-  RBF 1.0170431268334987
-  RMTC 1.5483147980515946
-  RMTB 1.0338582188178558
+  KPLS 1.020019652489361
+  KPLSK 1.0164675172271431
+  RBF 1.017043126837439
+  RMTC 2.543248921247831
+  RMTB 1.0339677426389438
   IDW 1.0163947834915366
   Best expert = IDW
-  Kriging 1.540784607392825
+  Kriging 1.5292005264525583
   LS 1.521071480977923
   QP 1.5434778008314816
-  KPLS 1.5262467470920507
-  KPLSK 1.5329970520881877
-  RBF 1.493811640632744
-  RMTC 1.635517120028801
-  RMTB 1.5464033315936672
+  KPLS 1.528256347974996
+  KPLSK 1.551349724012314
+  RBF 1.4938116417082332
+  RMTC 1.6358251469448486
+  RMTB 1.5464033492648523
   IDW 1.4918365835244014
   Best expert = IDW
-  Kriging 1.2191013043112668
+  Kriging 1.21847918939917
   LS 1.2431081352364661
   QP 1.2627639638801327
-  KPLS 1.224255312632745
-  KPLSK 1.2261843159790853
-  RBF 1.2131038247145347
-  RMTC 1.3135910205091708
-  RMTB 1.2334527112535287
+  KPLS 1.2257746152347502
+  KPLSK 1.22126452233064
+  RBF 1.2131038246365657
+  RMTC 1.3133779111801205
+  RMTB 1.2334527111293516
   IDW 1.2140402570940896
   Best expert = RBF
-  Kriging 1.0910512967452126
-  LS 1.0939223922787613
+  Kriging 1.0910144369666008
+  LS 1.093922392278761
   QP 1.0998924589863157
-  KPLS 1.0909047560427805
-  KPLSK 1.0910245734660722
-  RBF 1.0907867621379232
-  RMTC 1.0906652690338512
-  RMTB 1.0906812651891993
+  KPLS 1.0909047560424183
+  KPLSK 1.0910245734718471
+  RBF 1.0907867624306085
+  RMTC 1.0906652840505076
+  RMTB 1.0906812702308912
   IDW 1.1284069206525673
   Best expert = RMTC
