+++ doc/_src_docs/surrogate_models/gekpls.rst
@@ -91,7 +91,7 @@ Usage
    Training
      
      Training ...
-     Training - done. Time (sec):  0.0156000
+     Training - done. Time (sec):  0.0156002
   
 .. figure:: gekpls_Test_test_gekpls.png
   :scale: 80 %
