+++ doc/_src_docs/surrogate_models/krg.rst
@@ -71,7 +71,7 @@ Usage
    Training
      
      Training ...
-     Training - done. Time (sec):  0.0155997
+     Training - done. Time (sec):  0.0156000
   ___________________________________________________________________________
      
    Evaluation
