+++ smt/applications/tests/test_ego.py
@@ -1,0 +1,125 @@
+"""
+Author: Remi Lafage <remi.lafage@onera.fr> and Nathalie Bartoli
+This package is distributed under New BSD license.
+"""
+import unittest
+import numpy as np
+from sys import argv
+
+from smt.applications import EGO
+from smt.utils.sm_test_case import SMTestCase
+from smt.problems import Branin, Rosenbrock
+from smt.sampling_methods import FullFactorial
+
+import matplotlib.pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+
+
+class TestEGO(SMTestCase):
+    """
+    Test class
+    """
+
+    plot = None
+
+    @staticmethod
+    def function_test_1d(x):
+        # function xsinx
+        x = np.reshape(x, (-1,))
+        y = np.zeros(x.shape)
+        y = (x - 3.5) * np.sin((x - 3.5) / (np.pi))
+        return y.reshape((-1, 1))
+
+    def test_function_test_1d(self):
+        ndim = 1
+        niter = 15
+        xlimits = np.array([[0.0, 25.0]])
+
+        criterion = "EI"
+
+        ego = EGO(niter=niter, criterion=criterion, ndoe=3, xlimits=xlimits)
+
+        x_opt, y_opt, _, _, _, _, _ = ego.optimize(fun=TestEGO.function_test_1d)
+        print(x_opt, y_opt)
+
+        self.assertAlmostEqual(18.9, float(x_opt), delta=1)
+        self.assertAlmostEqual(-15.1, float(y_opt), delta=1)
+
+    def test_rosenbrock_2D(self):
+        ndim = 2
+        niter = 30
+        fun = Rosenbrock(ndim=ndim)
+        xlimits = fun.xlimits
+        criterion = "UCB"  #'EI' or 'SBO' or 'UCB'
+
+        xdoe = FullFactorial(xlimits=xlimits)(10)
+        ego = EGO(xdoe=xdoe, niter=niter, criterion=criterion, xlimits=xlimits)
+
+        x_opt, y_opt, _, _, _, _, _ = ego.optimize(fun=fun)
+
+        self.assertTrue(np.allclose([[1, 1]], x_opt, rtol=0.5))
+        self.assertAlmostEqual(0.0, float(y_opt), delta=1)
+
+    def test_branin_2D(self):
+        ndim = 2
+        niter = 10
+        fun = Branin(ndim=ndim)
+        xlimits = fun.xlimits
+        criterion = "UCB"  #'EI' or 'SBO' or 'UCB'
+
+        xdoe = FullFactorial(xlimits=xlimits)(10)
+        ego = EGO(xdoe=xdoe, niter=niter, criterion=criterion, xlimits=xlimits)
+
+        x_opt, y_opt, _, _, _, _, _ = ego.optimize(fun=fun)
+
+        # 3 optimal points possible: [-pi,12.275], [pi, 12.275], [9.42478,2.475]
+        self.assertTrue(
+            np.allclose([[-3.14, 12.275]], x_opt, rtol=0.1)
+            or np.allclose([[3.14, 12.275]], x_opt, rtol=0.1)
+            or np.allclose([[9.42, 2.475]], x_opt, rtol=0.1)
+        )
+        self.assertAlmostEqual(0.39, float(y_opt), delta=1)
+
+    @staticmethod
+    def run_ego_example():
+        import numpy as np
+        import six
+        from smt.applications import EGO
+        from smt.sampling_methods import FullFactorial
+
+        import sklearn
+        import matplotlib.pyplot as plt
+        from matplotlib import colors
+        from mpl_toolkits.mplot3d import Axes3D
+
+        def function_test_1d(x):
+            # function xsinx
+            import numpy as np
+
+            x = np.reshape(x, (-1,))
+            y = np.zeros(x.shape)
+            y = (x - 3.5) * np.sin((x - 3.5) / (np.pi))
+            return y.reshape((-1, 1))
+
+        ndim = 1
+        niter = 15
+        xlimits = np.array([[0.0, 25.0]])
+
+        criterion = "UCB"  #'EI' or 'SBO' or 'UCB'
+
+        ego = EGO(niter=niter, criterion=criterion, ndoe=3, xlimits=xlimits)
+
+        x_opt, y_opt, _, _, _, _, _ = ego.optimize(fun=function_test_1d)
+        print(x_opt, y_opt)
+        # Check if the optimal point is Xopt=18.9, Yopt =-15.1
+
+
+if __name__ == "__main__":
+    if "--plot" in argv:
+        TestEGO.plot = True
+        argv.remove("--plot")
+    if "--example" in argv:
+        TestEGO.run_ego_example()
+        exit()
+    unittest.main()
+
